+++ src/qec/analysis/__init__.py
@@ -2,6 +2,7 @@
 
 from .api import *  # noqa: F401,F403
 from .spectral_diversity_memory import SpectralDiversityConfig, SpectralDiversityMemory
+from .spectral_landscape_memory import SpectralLandscapeMemory
 from .spectral_frustration import SpectralFrustrationAnalyzer
 from .spectral_signature import SpectralSignature
 from .trap_memory import TrapMemoryConfig, TrapSubspaceMemory
@@ -11,6 +12,7 @@ __all__ = [
     "SpectralDiversityMemory",
     "SpectralFrustrationAnalyzer",
     "SpectralSignature",
+    "SpectralLandscapeMemory",
     "TrapMemoryConfig",
     "TrapSubspaceMemory",
 ]